Added python script to build the Knowledge Graph.
After the file Collection.py is run (Step 1), we get a CSV with all data of research papers from OpenAlex. Now, in this file, we build a knowledge graph by processing the values from the CSV. Have extracted a lot of nodes and relations from the CSV into the Knowledge Graph. You require a Neo4J DB connection to run this.